+++ src/game/filemenu/filemenu_createfile.c
@@ -448,7 +448,7 @@ void filemenu_choose_name_handle_input(MenuPanel* menu) {
                     sfx_play_sound(SOUND_CREATE_FILE_CHANGE_CHARSET);
                     menu->state = FM_INPUT_CHARSET_B;
                     filemenu_set_selected(menu, menu->col, menu->row);
-                    set_window_update(WIN_FILES_INPUT_KEYBOARD, (s32)filemenu_update_change_layout);
+                    set_window_update(WIN_FILES_INPUT_KEYBOARD, (intptr_t)filemenu_update_change_layout);
                 }
                 break;
             case MSG_CHAR_MENU_USE_CHARSET_A:
@@ -456,7 +456,7 @@ void filemenu_choose_name_handle_input(MenuPanel* menu) {
                     sfx_play_sound(SOUND_CREATE_FILE_CHANGE_CHARSET);
                     menu->state = FM_INPUT_CHARSET_A;
                     filemenu_set_selected(menu, menu->col, menu->row);
-                    set_window_update(WIN_FILES_INPUT_KEYBOARD, (s32)filemenu_update_change_layout);
+                    set_window_update(WIN_FILES_INPUT_KEYBOARD, (intptr_t)filemenu_update_change_layout);
                 }
                 break;
             case MSG_CHAR_MENU_END:
@@ -471,7 +471,7 @@ void filemenu_choose_name_handle_input(MenuPanel* menu) {
                 }
 
                 sfx_play_sound(SOUND_MENU_NEXT);
-                set_window_update(WIN_FILES_CONFIRM_OPTIONS, (s32) filemenu_update_show_name_confirm);
+                set_window_update(WIN_FILES_CONFIRM_OPTIONS, (intptr_t)filemenu_update_show_name_confirm);
                 gWindows[WIN_FILES_CONFIRM_OPTIONS].pos.y = 121;
                 gWindows[WIN_FILES_CONFIRM_OPTIONS].width = 69;
                 gWindows[WIN_FILES_CONFIRM_OPTIONS].height = 44;
@@ -521,19 +521,19 @@ void filemenu_choose_name_handle_input(MenuPanel* menu) {
         filemenu_filename_pos--;
         if (filemenu_filename_pos < 0) {
             filemenu_filename_pos = 0;
-            set_window_update(WIN_FILES_INPUT_FIELD, (s32) &filemenu_update_hidden_name_input);
-            set_window_update(WIN_FILES_INPUT_KEYBOARD, (s32) &filemenu_update_hidden_name_input);
-            set_window_update(WIN_FILES_TITLE, (s32) &filemenu_update_show_with_rotation);
-            set_window_update(WIN_FILES_STEREO, (s32) &filemenu_update_show_with_rotation);
-            set_window_update(WIN_FILES_MONO, (s32) &filemenu_update_show_with_rotation);
-            set_window_update(WIN_FILES_OPTION_LEFT, (s32) &filemenu_update_show_with_rotation);
-            set_window_update(WIN_FILES_OPTION_CENTER, (s32) &filemenu_update_show_with_rotation);
-            set_window_update(WIN_FILES_OPTION_RIGHT, (s32) &filemenu_update_show_with_rotation);
-            set_window_update(WIN_FILES_OPTION_LEFT, (s32) &filemenu_update_show_with_rotation);
-            set_window_update(WIN_FILES_SLOT1_BODY, (s32) &filemenu_update_show_with_rotation);
-            set_window_update(WIN_FILES_SLOT2_BODY, (s32) &filemenu_update_show_with_rotation);
-            set_window_update(WIN_FILES_SLOT3_BODY, (s32) &filemenu_update_show_with_rotation);
-            set_window_update(WIN_FILES_SLOT4_BODY, (s32) &filemenu_update_show_with_rotation);
+            set_window_update(WIN_FILES_INPUT_FIELD, (intptr_t)&filemenu_update_hidden_name_input);
+            set_window_update(WIN_FILES_INPUT_KEYBOARD, (intptr_t)&filemenu_update_hidden_name_input);
+            set_window_update(WIN_FILES_TITLE, (intptr_t)&filemenu_update_show_with_rotation);
+            set_window_update(WIN_FILES_STEREO, (intptr_t)&filemenu_update_show_with_rotation);
+            set_window_update(WIN_FILES_MONO, (intptr_t)&filemenu_update_show_with_rotation);
+            set_window_update(WIN_FILES_OPTION_LEFT, (intptr_t)&filemenu_update_show_with_rotation);
+            set_window_update(WIN_FILES_OPTION_CENTER, (intptr_t)&filemenu_update_show_with_rotation);
+            set_window_update(WIN_FILES_OPTION_RIGHT, (intptr_t)&filemenu_update_show_with_rotation);
+            set_window_update(WIN_FILES_OPTION_LEFT, (intptr_t)&filemenu_update_show_with_rotation);
+            set_window_update(WIN_FILES_SLOT1_BODY, (intptr_t)&filemenu_update_show_with_rotation);
+            set_window_update(WIN_FILES_SLOT2_BODY, (intptr_t)&filemenu_update_show_with_rotation);
+            set_window_update(WIN_FILES_SLOT3_BODY, (intptr_t)&filemenu_update_show_with_rotation);
+            set_window_update(WIN_FILES_SLOT4_BODY, (intptr_t)&filemenu_update_show_with_rotation);
             filemenu_currentMenu = FILE_MENU_MAIN;
             return;
         }
@@ -558,7 +558,7 @@ void filemenu_choose_name_handle_input(MenuPanel* menu) {
         }
 
         sfx_play_sound(SOUND_MENU_NEXT);
-        set_window_update(WIN_FILES_CONFIRM_OPTIONS, (s32) filemenu_update_show_name_confirm);
+        set_window_update(WIN_FILES_CONFIRM_OPTIONS, (intptr_t)filemenu_update_show_name_confirm);
 
         gWindows[WIN_FILES_CONFIRM_OPTIONS].pos.y = 121;
         gWindows[WIN_FILES_CONFIRM_OPTIONS].width = 69;
