+++ src/game/world/script_api/push_blocks.c
@@ -410,7 +410,7 @@ API_CALLABLE(SetPushBlock) {
         s32 posY = blockGrid->centerPos.y;
         s32 posZ = blockGrid->centerPos.z + (gridZ * BLOCK_GRID_SIZE) + (BLOCK_GRID_SIZE / 2);
         blockEntityID = create_entity(&Entity_PushBlock, posX, posY, posZ, 0, 0, 0, 0, MAKE_ENTITY_END);
-        bind_trigger_1(&EVS_PushWall_PushBlock, TRIGGER_WALL_PUSH, blockEntityID + EVT_ENTITY_ID_BIT, (s32)blockGrid, blockEntityID, 3);
+        bind_trigger_1(&EVS_PushWall_PushBlock, TRIGGER_WALL_PUSH, blockEntityID + EVT_ENTITY_ID_BIT, (intptr_t)blockGrid, blockEntityID, 3);
         script->varTable[0] = blockEntityID;
     }
